Data.Functor.MaybeLike

Description

Type constructors which can be turned into Maybe.

Documentation

class MaybeLike f where

This class may be defined in two ways:

Type constructors which can be turned into Maybe generically.

Type constructors which admit a natural transformation to Maybe.

We expect the following rules to hold for those instances of MaybeLike which are also Alternative:

likeMaybe empty = empty = Nothing

likeMaybe (x | y) = likeMaybe x | likeMaybe y

Methods

likeMaybe :: f a -> Maybe a