Mu.Schema

Description

Definition and interpretation of schemas in the vein of Avro, Protocol Buffers, or JSON Schema.

Each Schema is made out of types (which in turn be records or enumerations). A value which obbeys such a schema is called a Term. Conversion between Haskell types and schema types is mediated by the type classes ToSchema and FromSchema.

Schema definition

type Schema typeName fieldName = SchemaB Type typeName fieldName

A set of type definitions. In general, we can use any kind we want for both type and field names, although in practice you always want to use Symbol.

type Schema' = Schema Symbol Symbol

A set of type definitions, where the names of types and fields are defined by type-level strings (Symbols).

class KnownName (a :: k) where

Type names and field names can be of any kind, but for many uses we need a way to turn them into strings at run-time. This class generalizes KnownSymbol.

Methods

nameVal :: proxy a -> String

type TypeDef = TypeDefB Type

Defines a type in a schema. Each type can be: * a record: a list of key-value pairs, * an enumeration: an element of a list of choices, * a reference to a primitive type.

data TypeDefB builtin typeName fieldName

Defines a type in a schema, parametric on type representations.

Constructors

DRecord typeName [FieldDefB builtin typeName fieldName]
A list of key-value pairs.

DEnum typeName [ChoiceDef fieldName]
An element of a list of choices.

DSimple (FieldTypeB builtin typeName)
A reference to a primitive type.

newtype ChoiceDef fieldName

Defines each of the choices in an enumeration.

Constructors

ChoiceDef fieldName
One single choice from an enumeration.

type FieldDef = FieldDefB Type

Defines a field in a record by a name and the corresponding type.

data FieldDefB builtin typeName fieldName

Defines a field in a record, parametric on type representations.

Constructors

FieldDef fieldName (FieldTypeB builtin typeName)
One single field in a record.

type FieldType = FieldTypeB Type

Types of fields of a record. References to other types in the same schema are done via the TSchematic constructor.

data FieldTypeB builtin typeName

Types of fields of a record, parametric on type representations.

Constructors

TNull
Null, as found in Avro.

TPrimitive builtin
Reference to a primitive type, such as integers or Booleans. The set of supported primitive types depends on the protocol.

TSchematic typeName
Reference to another type in the schema.

TOption (FieldTypeB builtin typeName)
Optional value.

TList (FieldTypeB builtin typeName)
List of values.

TMap (FieldTypeB builtin typeName) (FieldTypeB builtin typeName)
Map of values. The set of supported key types depends on the protocol.

TUnion [FieldTypeB builtin typeName]
Represents a choice between types.

Lookup type in schema

type family (sch :: Schema t f) :/: (name :: t) :: TypeDef t f where ...

Lookup a type in a schema by its name.

Equations

'[] :/: name = TypeError (('Text "Cannot find type " :<>: 'ShowType name) :<>: 'Text " in the schema")

('DRecord name fields ': rest) :/: name = 'DRecord name fields

('DEnum name choices ': rest) :/: name = 'DEnum name choices

(other ': rest) :/: name = rest :/: name

Interpretation of schemas

data Term (sch :: Schema typeName fieldName) (t :: TypeDef typeName fieldName) where

Interpretation of a type in a schema.

Constructors

TRecord :: NP (Field sch) args -> Term sch ('DRecord name args)
A record given by the value of its fields.

TEnum :: NS Proxy choices -> Term sch ('DEnum name choices)
An enumeration given by one choice.

TSimple :: FieldValue sch t -> Term sch ('DSimple t)
A primitive value.

data Field (sch :: Schema typeName fieldName) (f :: FieldDef typeName fieldName) where

Interpretation of a field.

Constructors

Field :: FieldValue sch t -> Field sch ('FieldDef name t)
A single field. Note that the contents are wrapped in a w type constructor.

data FieldValue (sch :: Schema typeName fieldName) (t :: FieldType typeName) where

Interpretation of a field type, by giving a value of that type.

Constructors

FNull :: FieldValue sch 'TNull
Null value, as found in Avro and JSON.

FPrimitive :: t -> FieldValue sch ('TPrimitive t)
Value of a primitive type.

FSchematic :: Term sch (sch :/: t) -> FieldValue sch ('TSchematic t)
Term of another type in the schema.

FOption :: Maybe (FieldValue sch t) -> FieldValue sch ('TOption t)
Optional value.

FList :: [FieldValue sch t] -> FieldValue sch ('TList t)
List of values.

FMap :: Ord (FieldValue sch k) => Map (FieldValue sch k) (FieldValue sch v) -> FieldValue sch ('TMap k v)
Dictionary (key-value map) of values.

FUnion :: NS (FieldValue sch) choices -> FieldValue sch ('TUnion choices)
One single value of one of the specified types.

data NS (a :: k -> Type) (b :: [k]) where

Constructors

Z :: forall k (a :: k -> Type) (x :: k) (xs :: [k]). a x -> NS a (x ': xs)

S :: forall k (a :: k -> Type) (xs :: [k]) (x :: k). NS a xs -> NS a (x ': xs)

data NP (a :: k -> Type) (b :: [k]) where

Constructors

Nil :: forall k (a :: k -> Type). NP a ('[] :: [k])

(:*) :: forall k (a :: k -> Type) (x :: k) (xs :: [k]). a x -> NP a xs -> NP a (x ': xs)

data Proxy (t :: k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy

Conversion from types to schemas

newtype WithSchema (sch :: Schema tn fn) (sty :: tn) a where

Tags a value with its schema. For usage with deriving via.

Constructors

WithSchema :: forall tn fn (sch :: Schema tn fn) (sty :: tn) a. a -> WithSchema sch sty a

unWithSchema :: forall tn fn (sch :: Schema tn fn) (sty :: tn) a. WithSchema sch sty a -> a

Accessor for WithSchema. Intended for usage with TypeApplications.

class FromSchema (sch :: Schema typeName fieldName) (sty :: typeName) (t :: Type) | sch t -> sty where

Defines the conversion from a Term which follows the schema sch into a type t. You can give an optional mapping between the field names of t and that of sty by means of CustomFieldMapping.

Minimal complete definition

Nothing

Methods

fromSchema :: Term sch (sch :/: sty) -> t

Conversion from schema term to Haskell type.

default fromSchema :: (Generic t, GFromSchemaTypeDef sch '[] (sch :/: sty) (Rep t)) => Term sch (sch :/: sty) -> t

fromSchema' :: forall fn tn (sch :: Schema tn fn) t sty. FromSchema sch sty t => Term sch (sch :/: sty) -> t

Conversion from schema term to Haskell type. This version is intended for usage with TypeApplications: > fromSchema' @MySchema mySchemaTerm

class ToSchema (sch :: Schema typeName fieldName) (sty :: typeName) (t :: Type) | sch t -> sty where

Defines the conversion of a type t into a Term which follows the schema sch. You can give an optional mapping between the field names of t and that of sty by means of CustomFieldMapping.

Minimal complete definition

Nothing

Methods

toSchema :: t -> Term sch (sch :/: sty)

Conversion from Haskell type to schema term.

default toSchema :: (Generic t, GToSchemaTypeDef sch '[] (sch :/: sty) (Rep t)) => t -> Term sch (sch :/: sty)

toSchema' :: forall fn tn (sch :: Schema tn fn) t sty. ToSchema sch sty t => t -> Term sch (sch :/: sty)

Conversion from Haskell type to schema term. This version is intended for usage with TypeApplications: > toSchema' @MySchema myValue

newtype CustomFieldMapping (sty :: typeName) (fmap :: [Mapping Symbol fieldName]) a

By default, the names of the fields in the Haskell type and those of the schema types must coincide. By using this wrapper you can override this default setting.

This type should be used with DerivingVia, as follows:

type MyCustomFieldMapping = '[ "A" ':-> "a", ...]
data MyHaskellType = ...
  deriving ( ToSchema   f MySchema "MySchemaType" MyHaskellType
           , FromSchema f MySchema "MySchemaType" MyHaskellType)
    via (CustomFieldMapping "MySchemaType" MyCustomFieldMapping MyHaskellType)

Constructors

CustomFieldMapping a

newtype Underlying basic logical

This 'newtype' is used to wrap types for which we want a "logical" representation as a Haskell type, but the underlying representation is lower level, like UUIDs as ByteStrings.

Constructors

Underlying

Fields

unUnderlying :: logical

class UnderlyingConversion basic logical where

This class defines the actual conversion between a "logical" type and its low-level representation.

Methods

toUnderlying :: logical -> basic

fromUnderlying :: basic -> logical

Mappings between fields

data Mapping a b

Defines a mapping between two elements.

Constructors

a :-> b

type Mappings a b = [Mapping a b]

Defines a set of mappings between elements of a and b.

type family MappingRight (ms :: Mappings a b) (v :: a) :: b where ...

Finds the corresponding right value of v in a mapping ms. When the kinds are Symbol, return the same value if not found. When the return type is Type, return ' ()' if the value is not found.

Equations

MappingRight '[] (v :: Symbol) = v :: Symbol

MappingRight '[] (v :: Symbol) = () :: Type

MappingRight '[] v = TypeError ('Text "Cannot find value " :<>: 'ShowType v)

MappingRight ((x :-> y) ': rest) x = y

MappingRight (other ': rest) x = MappingRight rest x

type family MappingLeft (ms :: Mappings a b) (v :: b) :: a where ...

Finds the corresponding left value of v in a mapping ms. When the kinds are Symbol, return the same value if not found. When the return type is Type, return ' ()' if the value is not found.

Equations

MappingLeft '[] (v :: Symbol) = v :: Symbol

MappingLeft '[] (v :: Symbol) = () :: Type

MappingLeft '[] v = TypeError ('Text "Cannot find value " :<>: 'ShowType v)

MappingLeft ((x :-> y) ': rest) y = x

MappingLeft (other ': rest) y = MappingLeft rest y

Field annotations

type family AnnotatedSchema domain (sch :: Schema typeName fieldName) :: [Annotation domain typeName fieldName]

This type family links each schema to its corresponding annotations from one domain.

type AnnotationDomain = Type

Each annotation belongs to a domain.

data Annotation domain typeName fieldName where

Annotations proper.

Constructors

AnnSchema :: domain -> Annotation domain typeName fieldName
Annotation over the whole schema.

AnnType :: typeName -> domain -> Annotation domain typeName fieldName
Annotation over a type in the schema.

AnnField :: typeName -> fieldName -> domain -> Annotation domain typeName fieldName
Annotation over a field in a record or a choice in an enumeration.