Mu.Schema.Class

Description

This module defines a couple of type classes ToSchema and FromSchema to turn Haskell types back and forth mu-haskell Terms.

In most cases, the instances can be automatically derived. If you enable the extensions DeriveGeneric and DeriveAnyClass, you can do:

data MyHaskellType = ...
  deriving ( ToSchema   f MySchema "MySchemaType" MyHaskellType
           , FromSchema f MySchema "MySchemaType" MyHaskellType)

If the default mapping which required identical names for fields in the Haskell and schema types does not suit you, use CustomFieldMapping.

Documentation

newtype WithSchema (w :: Type -> Type) (sch :: Schema tn fn) (sty :: tn) a

Tags a value with its schema. For usage with deriving via.

Constructors

WithSchema a

class FromSchema (w :: Type -> Type) (sch :: Schema typeName fieldName) (sty :: typeName) (t :: Type) | sch t -> sty where

Defines the conversion from a Term which follows the schema sch into a type t. You can give an optional mapping between the field names of t and that of sty by means of CustomFieldMapping.

Minimal complete definition

Nothing

Methods

fromSchema :: Term w sch (sch :/: sty) -> t

Conversion from schema term to Haskell type.

fromSchema :: (Generic t, GFromSchemaTypeDef w sch '[] (sch :/: sty) (Rep t)) => Term w sch (sch :/: sty) -> t

Conversion from schema term to Haskell type.

fromSchema' :: forall fn tn (sch :: Schema tn fn) w t sty. FromSchema w sch sty t => Term w sch (sch :/: sty) -> t

Conversion from schema term to Haskell type. This version is intended for usage with TypeApplications: > fromSchema' @MySchema mySchemaTerm

class ToSchema (w :: Type -> Type) (sch :: Schema typeName fieldName) (sty :: typeName) (t :: Type) | sch t -> sty where

Defines the conversion of a type t into a Term which follows the schema sch. You can give an optional mapping between the field names of t and that of sty by means of CustomFieldMapping.

Minimal complete definition

Nothing

Methods

toSchema :: t -> Term w sch (sch :/: sty)

Conversion from Haskell type to schema term.

toSchema :: (Generic t, GToSchemaTypeDef w sch '[] (sch :/: sty) (Rep t)) => t -> Term w sch (sch :/: sty)

Conversion from Haskell type to schema term.

toSchema' :: forall fn tn (sch :: Schema tn fn) w t sty. ToSchema w sch sty t => t -> Term w sch (sch :/: sty)

Conversion from Haskell type to schema term. This version is intended for usage with TypeApplications: > toSchema' @MySchema myValue

newtype CustomFieldMapping (sty :: typeName) (fmap :: [Mapping Symbol fieldName]) a

By default, the names of the fields in the Haskell type and those of the schema types must coincide. By using this wrapper you can override this default setting.

This type should be used with DerivingVia, as follows:

type MyCustomFieldMapping = '[ "A" ':-> "a", ...]
data MyHaskellType = ...
  deriving ( ToSchema   f MySchema "MySchemaType" MyHaskellType
           , FromSchema f MySchema "MySchemaType" MyHaskellType)
    via (CustomFieldMapping "MySchemaType" MyCustomFieldMapping MyHaskellType)

Constructors

CustomFieldMapping a

data Mapping a b

Defines a mapping between two elements.

Constructors

a :-> b

type Mappings a b = [Mapping a b]

Defines a set of mappings between elements of a and b.

type family MappingRight (ms :: Mappings a b) (v :: a) :: b where ...

Finds the corresponding right value of v in a mapping ms. When the kinds are Symbol, return the same value if not found.

Equations

MappingRight '[] (v :: Symbol) = v

MappingRight '[] v = TypeError (Text "Cannot find value " :<>: ShowType v)

MappingRight ((x :-> y) ': rest) x = y

MappingRight (other ': rest) x = MappingRight rest x

type family MappingLeft (ms :: Mappings a b) (v :: b) :: a where ...

Finds the corresponding left value of v in a mapping ms. When the kinds are Symbol, return the same value if not found.

Equations

MappingLeft '[] (v :: Symbol) = v

MappingLeft '[] v = TypeError (Text "Cannot find value " :<>: ShowType v)

MappingLeft ((x :-> y) ': rest) y = x

MappingLeft (other ': rest) y = MappingLeft rest y

transSchema :: forall fn tn (sch :: Schema tn fn) sty u v a b. (ToSchema u sch sty a, FromSchema v sch sty b, Functor u, forall k. Ord (FieldValue u sch k) => Ord (FieldValue v sch k)) => (forall x. u x -> v x) -> Proxy sch -> a -> b

Changes the underlying wrapper of a Haskell type, by converting back and forth Terms with those wrappers.

Internal use only

class GToSchemaRecord (w :: * -> *) (sch :: Schema ts fs) (fmap :: Mappings Symbol fs) (args :: [FieldDef ts fs]) (f :: * -> *) where

For internal use only: generic conversion of a list of fields.

Methods

toSchemaRecord :: Proxy fmap -> f a -> NP (Field w sch) args