Mu.Schema.Conversion.SchemaToTypes

Description

This module is deprecated. Haskell types corresponding to schema types should be written manually.

Documentation

generateTypesFromSchema :: Namer -> Name -> Q [Dec]

Generates types to represent each of the types in a given schema. You should call it as: > $(generateTypesFromSchema f 'Schema) where f is a function String -> String which obtains the Haskell name for a type given the name in the schema. The second argument is simply the name of the schema.

type Namer = String -> String

Generate the name from each new Haskell type from the name given in the schema.