Mu.Schema.Interpretation

Description

This module defines Terms which comply with a given Schema. These Terms are the main form of values used internally by mu-haskell.

In this module we make use of NP and NS as defined by sop-core. These are the n-ary versions of a pair and Either, respectively. In other words, NP puts together a bunch of values of different types, NS allows you to choose from a bunch of types.

Interpretation

data Term (sch :: Schema typeName fieldName) (t :: TypeDef typeName fieldName) where

Interpretation of a type in a schema.

Constructors

TRecord :: NP (Field sch) args -> Term sch ('DRecord name args)
A record given by the value of its fields.

TEnum :: NS Proxy choices -> Term sch ('DEnum name choices)
An enumeration given by one choice.

TSimple :: FieldValue sch t -> Term sch ('DSimple t)
A primitive value.

data Field (sch :: Schema typeName fieldName) (f :: FieldDef typeName fieldName) where

Interpretation of a field.

Constructors

Field :: FieldValue sch t -> Field sch ('FieldDef name t)
A single field. Note that the contents are wrapped in a w type constructor.

data FieldValue (sch :: Schema typeName fieldName) (t :: FieldType typeName) where

Interpretation of a field type, by giving a value of that type.

Constructors

FNull :: FieldValue sch 'TNull
Null value, as found in Avro and JSON.

FPrimitive :: t -> FieldValue sch ('TPrimitive t)
Value of a primitive type.

FSchematic :: Term sch (sch :/: t) -> FieldValue sch ('TSchematic t)
Term of another type in the schema.

FOption :: Maybe (FieldValue sch t) -> FieldValue sch ('TOption t)
Optional value.

FList :: [FieldValue sch t] -> FieldValue sch ('TList t)
List of values.

FMap :: Ord (FieldValue sch k) => Map (FieldValue sch k) (FieldValue sch v) -> FieldValue sch ('TMap k v)
Dictionary (key-value map) of values.

FUnion :: NS (FieldValue sch) choices -> FieldValue sch ('TUnion choices)
One single value of one of the specified types.

data NS (a :: k -> Type) (b :: [k]) where

Constructors

Z :: forall k (a :: k -> Type) (x :: k) (xs :: [k]). a x -> NS a (x ': xs)

S :: forall k (a :: k -> Type) (xs :: [k]) (x :: k). NS a xs -> NS a (x ': xs)

data NP (a :: k -> Type) (b :: [k]) where

Constructors

Nil :: forall k (a :: k -> Type). NP a ('[] :: [k])

(:*) :: forall k (a :: k -> Type) (x :: k) (xs :: [k]). a x -> NP a xs -> NP a (x ': xs)

data Proxy (t :: k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy