Mu.Schema.Interpretation.Anonymous

Description

This module provides "anonymous terms". These terms can be used when you don't want to write your own Haskell type, but simply have a quick and dirty interpretation for a schema type. An important limitation is that anonymous terms may only contain primitive fields.

The names of the types exposed in this module refer to the amount of fields in the record. Hence, use V0 for empty record, V1 for a record with one field, V2 for two, and so forth.

Documentation

data V0 w sch sty where

Anonymous term for a record with zero fields.

Constructors

V0 :: (sch :/: sty) ~ DRecord nm '[] => V0 w sch sty

data V1 w sch sty where

Anonymous term for a record with one field.

Constructors

V1 :: (sch :/: sty) ~ DRecord nm '[FieldDef f (TPrimitive a)] => w a -> V1 w sch sty

data V2 w sch sty where

Anonymous term for a record with two fields.

Constructors

V2 :: (sch :/: sty) ~ DRecord nm '[FieldDef f (TPrimitive a), FieldDef g (TPrimitive b)] => w a -> w b -> V2 w sch sty