Mu.Schema.Registry

Description

A registry of schemas saves the different schemas supported by an application. Since messages and protocols may evolve, it's useful to keep an updated view of the different shapes of data we can handle.

Examples of registries are found in Kafka and Compendium.

Registry of schemas

type Registry = Mappings Nat Schema'

A Registry is defined as a map from version numbers to type-level schemas.

Implementation note: you must write newer schemas at the head of the Registry. Otherwise, older schemas take precedence during conversion.

fromRegistry :: forall r t. FromRegistry r t => Term -> Maybe t

Converts a schemaless term into a value by checking all the possible schemas in a Registry.

Implementation note: schemas are checked in the same order in which they appear in the Registry definition.

Terms without an associated schema

data Term where

Interpretation of a type in a schema.

Constructors

TRecord :: [Field] -> Term
A record given by the value of its fields.

TEnum :: Int -> Term
An enumeration given by one choice.

TSimple :: FieldValue -> Term
A primitive value.

data Field where

Interpretation of a field.

Constructors

Field :: Text -> FieldValue -> Field
A single field given by its name and its value.

data FieldValue where

Interpretation of a field type, by giving a value of that type.

Constructors

FNull :: FieldValue

FPrimitive :: (Typeable t, Eq t, Ord t, Show t) => t -> FieldValue

FSchematic :: Term -> FieldValue

FOption :: Maybe FieldValue -> FieldValue

FList :: [FieldValue] -> FieldValue

FMap :: Map FieldValue FieldValue -> FieldValue